[Feat]#126 - 기능 완성
